fix(parser): close grammar gaps blocking Standard-Library CoordSys.ili
Adds FORMAT BASED ON, optional PARAMETER blocks in class/structure
bodies, reserved keywords (REFSYSTEM/COORDSYSTEM) as qualified-name
tail segments, NUMERIC with suffixes without range, and `+` signed
numbers. CoordSys.ili + Units.ili shipped as snapshot fixtures with
provenance README. 138/138 tests green.
